feat(AutoSequence.java): use auto sequences to plan out autonimous op modes
The abstract auto sequence classes, are used to define autonomous modes. The idea is to utilize Sequential Command Groups to plan out the autonimous phase. Since command groups can be nested, the auto can be divided up into a combination of Parallel and Sequential Command Groups, which will run in sequence.